Extract logger related methods to a dedicated file.
