+++ Core/CoreTests/Analytics/UIViewControllerLoggingTests.swift
@@ -1,0 +1,36 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2024-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+import Core
+import XCTest
+
+class UIViewControllerLoggingTests: CoreTestCase {
+
+    func testAnalyticsClassName() {
+        let courseListView = CoreHostingController(PandaGallery())
+
+        XCTAssertEqual((nil as UIViewController?).loggableName, "unknown")
+        XCTAssertEqual(ProfileSettingsViewController().loggableName, "ProfileSettingsViewController")
+        XCTAssertEqual(courseListView.loggableName, "PandaGallery")
+        XCTAssertEqual(UINavigationController(rootViewController: courseListView).loggableName, "PandaGallery")
+
+        let splitView = UISplitViewController()
+        splitView.viewControllers = [UINavigationController(rootViewController: courseListView)]
+        XCTAssertEqual(splitView.loggableName, "PandaGallery")
+    }
+}
